third_party/WebKit/LayoutTests/media/track/vtt-cue-exceptions-expected.txt

Verify that VTTCue exceptions are properly messaged to developers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cue.position = -1; threw exception IndexSizeError: Failed to set the 'position' property on 'VTTCue': The value provided (-1) is outside the range [0, 100]..
PASS cue.position = 101; threw exception IndexSizeError: Failed to set the 'position' property on 'VTTCue': The value provided (101) is outside the range [0, 100]..
PASS cue.size = -1; threw exception IndexSizeError: Failed to set the 'size' property on 'VTTCue': The value provided (-1) is outside the range [0, 100]..
PASS cue.size = 101; threw exception IndexSizeError: Failed to set the 'size' property on 'VTTCue': The value provided (101) is outside the range [0, 100]..
PASS successfullyParsed is true

TEST COMPLETE
